requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

torch>=2.0.0
transformers>=4.41.0
sentencepiece>=0.1.99
sentence-transformers>=2.2.2

scikit-learn>=1.3.0
numpy>=1.24.0

shap>=0.44.0
pinecone>=5.0.0
faiss-cpu>=1.7.4

langdetect>=1.0.9
python-dotenv>=1.0.0

fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
python-multipart==0.0.9
